Fix pluginization UI cache and memory disabled state
