+++ frontend/src/components/chat/ChatInput.tsx
@@ -213,7 +213,8 @@ export const ChatInput = memo(function ChatInput({
   const pluginShortcutOptions = chatInputOptions.filter(
     (option) => option.shortcut && option.panel && pluginShortcutPanelIds.has(option.panel),
   );
-  const mentionMode = activePluginMentionProvider?.mode ?? "persona";
+  const mentionMode: "team" | "persona" =
+    activePluginMentionProvider?.mode === "team" ? "team" : "persona";
   const mentionEnabled =
     activePluginMentionProvider ? true : !!onUsePersonaPreset;
