+++ frontend/src/components/panels/PluginRuntimePanel.tsx
@@ -25,12 +25,16 @@ import { useTranslation } from "react-i18next";
 import { Button, IconButton } from "../common";
 import { EmptyState } from "../common/EmptyState";
 import { PanelHeader } from "../common/PanelHeader";
-import {
-  buildPluginContributionPreview,
-  type PluginContributionSnapshot,
-} from "../../extensions/coreContributions";
 import { useAuth } from "../../hooks/useAuth";
 import { usePluginRuntime } from "../../hooks/usePluginRuntime";
+import { buildPluginRuntimeImpactSummary } from "./pluginRuntimeImpactSummary";
+import {
+  legacyFrontendContributionCount,
+  pluginContributionGroups,
+  pluginContributionLabels,
+  structuredFrontendContributionCount,
+  type PluginContributionGroup,
+} from "./pluginRuntimePanelUtils";
 import {
   Permission,
   type ArchivedPluginPackage,
@@ -55,306 +59,6 @@ function formatCounts(values: Record<string, number>): string {
   const entries = Object.entries(values);
   if (entries.length === 0) return "0";
   return entries.map(([key, value]) => `${key}: ${value}`).join(" · ");
-}
-
-function formatToolLabel(tool: PluginRuntimePlugin["tools"][number]): string {
-  if (tool.legacy_ids.length === 0) return tool.name;
-  return `${tool.name} (${tool.legacy_ids.join(" / ")})`;
-}
-
-function formatChatInputOptionLabel(
-  option: PluginRuntimePlugin["frontend"]["chat_input_options"][number],
-  prefix: string,
-): string {
-  const effects = option.suppresses_core_persona_selector
-    ? " suppresses core persona selector"
-    : "";
-  return `${prefix} ${option.id}${effects}`;
-}
-
-function contributionId(value: string | { id: string }): string {
-  return typeof value === "string" ? value : value.id;
-}
-
-function formatToolRendererContribution(
-  value: PluginRuntimePlugin["frontend"]["tool_renderers"][number],
-): string {
-  const tools = typeof value === "string" ? [] : value.tool_names ?? [];
-  return tools.length > 0
-    ? `${contributionId(value)} (${tools.join(" / ")})`
-    : contributionId(value);
-}
-
-function formatFileViewerContribution(
-  value: PluginRuntimePlugin["frontend"]["file_viewers"][number],
-): string {
-  const extensions = typeof value === "string" ? [] : value.extensions ?? [];
-  return extensions.length > 0
-    ? `${contributionId(value)} (${extensions.join(" / ")})`
-    : contributionId(value);
-}
-
-function formatSkillImporterContribution(
-  value: PluginRuntimePlugin["frontend"]["skill_importers"][number],
-): string {
-  return typeof value === "string" ? value : `${value.id} (${value.source})`;
-}
-
-function formatChannelConnectorContribution(
-  value: PluginRuntimePlugin["frontend"]["channel_connectors"][number],
-): string {
-  return typeof value === "string" ? value : `${value.id} (${value.channel_type})`;
-}
-
-function uniqueValues(values: readonly string[]): string[] {
-  return Array.from(new Set(values.filter(Boolean)));
-}
-
-function legacyFrontendDeclarationLabels(plugin: PluginRuntimePlugin): string[] {
-  return uniqueValues([
-    ...plugin.frontend.routes.map((value) => `frontend route ${value}`),
-    ...plugin.frontend.panels.map((value) => `panel ${value}`),
-    ...plugin.frontend.nav_items.map((value) => `nav ${value}`),
-  ]);
-}
-
-function structuredFrontendDeclarationLabels(plugin: PluginRuntimePlugin): string[] {
-  const assetSlots = plugin.package?.frontend_assets?.slots ?? [];
-  return uniqueValues([
-    ...plugin.frontend.app_tabs.map((value) => `app tab ${value.path || value.tab}`),
-    ...plugin.frontend.app_panels.map((value) => `app panel ${value.renderer}`),
-    ...plugin.frontend.sidebar_items.map((value) => `sidebar ${value.path}`),
-    ...plugin.frontend.user_menu_items.map((value) => `user menu ${value.path}`),
-    ...plugin.frontend.chat_input_options.map((value) =>
-      formatChatInputOptionLabel(value, "chat option"),
-    ),
-    ...plugin.frontend.chat_input_panels.map((value) => `chat panel ${value.renderer}`),
-    ...plugin.frontend.mention_providers.map((value) => `mention ${value.mode}`),
-    ...plugin.frontend.welcome_surfaces.map((value) => `welcome surface ${value.renderer}`),
-    ...plugin.frontend.project_options.map((value) => `project option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.session_options.map((value) => `session option ${plugin.plugin_id}.${value.key}`),
-    ...(plugin.frontend.channel_options ?? []).map((value) => `channel option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.tool_renderers.map((value) => `renderer ${formatToolRendererContribution(value)}`),
-    ...plugin.frontend.file_viewers.map((value) => `viewer ${formatFileViewerContribution(value)}`),
-    ...plugin.frontend.skill_importers.map((value) => `importer ${formatSkillImporterContribution(value)}`),
-    ...plugin.frontend.channel_connectors.map((value) => `connector ${formatChannelConnectorContribution(value)}`),
-    ...plugin.frontend.message_actions.map((value) => `message action ${value.id}`),
-    ...plugin.frontend.assistant_identity_resolvers.map((value) => `assistant identity ${value.resolver}`),
-    ...plugin.frontend.agent_categories.map((value) => `agent category ${value.id}`),
-    ...plugin.frontend.scheduled_task_options.map((value) => `scheduled task option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.settings_sections.map((value) => `settings ${value}`),
-    ...plugin.frontend.i18n_namespaces.map((value) => `i18n ${value}`),
-    ...assetSlots.map((value) => `asset slot ${value}`),
-  ]);
-}
-
-function frontendDeclarationLabels(plugin: PluginRuntimePlugin): string[] {
-  return uniqueValues([
-    ...structuredFrontendDeclarationLabels(plugin),
-    ...legacyFrontendDeclarationLabels(plugin),
-  ]);
-}
-
-function structuredFrontendContributionCount(plugin: PluginRuntimePlugin): number {
-  return (
-    plugin.frontend.app_tabs.length +
-    plugin.frontend.app_panels.length +
-    plugin.frontend.sidebar_items.length +
-    plugin.frontend.user_menu_items.length +
-    plugin.frontend.chat_input_options.length +
-    plugin.frontend.chat_input_panels.length +
-    plugin.frontend.mention_providers.length +
-    plugin.frontend.welcome_surfaces.length +
-    plugin.frontend.project_options.length +
-    plugin.frontend.session_options.length +
-    (plugin.frontend.channel_options ?? []).length +
-    plugin.frontend.scheduled_task_options.length +
-    plugin.frontend.tool_renderers.length +
-    plugin.frontend.file_viewers.length +
-    plugin.frontend.skill_importers.length +
-    plugin.frontend.channel_connectors.length +
-    plugin.frontend.message_actions.length +
-    plugin.frontend.assistant_identity_resolvers.length +
-    plugin.frontend.agent_categories.length +
-    plugin.frontend.settings_sections.length +
-    plugin.frontend.i18n_namespaces.length +
-    (plugin.package?.frontend_assets?.slots.length ?? 0)
-  );
-}
-
-function legacyFrontendContributionCount(plugin: PluginRuntimePlugin): number {
-  return (
-    plugin.frontend.routes.length +
-    plugin.frontend.panels.length +
-    plugin.frontend.nav_items.length
-  );
-}
-
-function declaredRuntimeEntryLabels(plugin: PluginRuntimePlugin): string[] {
-  return uniqueValues([
-    ...plugin.routes.map((route) => `route ${route.prefix}`),
-    ...plugin.agents.map((agent) => `agent ${agent.id}`),
-    ...plugin.tools.map((tool) => `tool ${formatToolLabel(tool)}`),
-    ...frontendDeclarationLabels(plugin),
-  ]);
-}
-
-function resourceActionLabels(values: Record<string, number>): string[] {
-  return Object.entries(values).map(([action, count]) => `${action}: ${count}`);
-}
-
-function pluginContributionLabels(plugin: PluginRuntimePlugin): string[] {
-  return uniqueValues([
-    ...plugin.routes.map((route) => `API ${route.prefix}`),
-    ...plugin.agents.map((agent) => `Agent ${agent.id}`),
-    ...plugin.tools.map((tool) => `Tool ${formatToolLabel(tool)}`),
-    ...legacyFrontendDeclarationLabels(plugin).map((label) => `Legacy ${label}`),
-    ...plugin.frontend.app_tabs.map((value) => `App Tab ${value.path || value.tab}`),
-    ...plugin.frontend.app_panels.map((value) => `App Panel ${value.renderer}`),
-    ...plugin.frontend.sidebar_items.map((value) => `Sidebar ${value.path}`),
-    ...plugin.frontend.user_menu_items.map((value) => `User Menu ${value.path}`),
-    ...plugin.frontend.chat_input_options.map((value) =>
-      formatChatInputOptionLabel(value, "Chat Option"),
-    ),
-    ...plugin.frontend.chat_input_panels.map((value) => `Chat Panel ${value.renderer}`),
-    ...plugin.frontend.message_actions.map((value) => `Message Action ${value.id}`),
-    ...plugin.frontend.mention_providers.map((value) => `Mention ${value.mode}`),
-    ...plugin.frontend.welcome_surfaces.map((value) => `Welcome Surface ${value.renderer}`),
-    ...plugin.frontend.assistant_identity_resolvers.map((value) => `Assistant Identity ${value.resolver}`),
-    ...plugin.frontend.agent_categories.map((value) => `Agent Category ${value.id}`),
-    ...plugin.frontend.project_options.map((value) => `Project Option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.session_options.map((value) => `Session Option ${plugin.plugin_id}.${value.key}`),
-    ...(plugin.frontend.channel_options ?? []).map((value) => `Channel Option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.scheduled_task_options.map((value) => `Scheduled Task Option ${plugin.plugin_id}.${value.key}`),
-    ...plugin.frontend.tool_renderers.map((value) => `Renderer ${formatToolRendererContribution(value)}`),
-    ...plugin.frontend.file_viewers.map((value) => `Viewer ${formatFileViewerContribution(value)}`),
-    ...plugin.frontend.skill_importers.map((value) => `Importer ${formatSkillImporterContribution(value)}`),
-    ...plugin.frontend.channel_connectors.map((value) => `Connector ${formatChannelConnectorContribution(value)}`),
-    ...plugin.frontend.settings_sections.map((value) => `Settings ${value}`),
-    ...plugin.frontend.i18n_namespaces.map((value) => `I18n ${value}`),
-    ...(plugin.package?.frontend_assets?.slots ?? []).map((value) => `Asset Slot ${value}`),
-  ]);
-}
-
-interface PluginContributionGroup {
-  id: string;
-  title: string;
-  entries: string[];
-}
-
-function pluginContributionGroups(plugin: PluginRuntimePlugin): PluginContributionGroup[] {
-  const assetSlots = plugin.package?.frontend_assets?.slots ?? [];
-  const groups: PluginContributionGroup[] = [
-    {
-      id: "backend",
-      title: "Backend",
-      entries: uniqueValues([
-        ...plugin.routes.map((route) => `API ${route.prefix} -> ${route.module}`),
-        ...plugin.tools.map((tool) => `Tool ${formatToolLabel(tool)} -> ${tool.module}`),
-        ...plugin.permissions.map((permission) => `Permission ${permission}`),
-      ]),
-    },
-    {
-      id: "app-ui",
-      title: "App UI",
-      entries: uniqueValues([
-        ...plugin.frontend.app_tabs.map((value) => `Tab ${value.path || value.tab}`),
-        ...plugin.frontend.app_panels.map((value) => `Panel ${value.renderer}`),
-        ...plugin.frontend.sidebar_items.map((value) => `Sidebar ${value.path}`),
-        ...plugin.frontend.user_menu_items.map((value) => `User Menu ${value.path}`),
-        ...plugin.frontend.message_actions.map((value) => `Message Action ${value.id}`),
-      ]),
-    },
-    {
-      id: "chat-ui",
-      title: "Chat UI",
-      entries: uniqueValues([
-        ...plugin.frontend.chat_input_options.map((value) =>
-          formatChatInputOptionLabel(value, "Chat Option"),
-        ),
-        ...plugin.frontend.chat_input_panels.map((value) => `Chat Panel ${value.renderer}`),
-        ...plugin.frontend.mention_providers.map((value) => `Mention ${value.mode} -> ${value.provider}`),
-        ...plugin.frontend.welcome_surfaces.map((value) => `Welcome ${value.agent_id} -> ${value.renderer}`),
-        ...plugin.frontend.assistant_identity_resolvers.map((value) => `Assistant Identity ${value.agent_id} -> ${value.resolver}`),
-      ]),
-    },
-    {
-      id: "agent",
-      title: "Agent",
-      entries: uniqueValues([
-        ...plugin.agents.map((agent) => `Agent ${agent.id} -> ${agent.module}`),
-        ...plugin.frontend.agent_categories.map((value) => `Category ${value.id}`),
-      ]),
-    },
-    {
-      id: "scoped-options",
-      title: "Scoped Options",
-      entries: uniqueValues([
-        ...plugin.frontend.project_options.map((value) => `Project ${plugin.plugin_id}.${value.key}`),
-        ...plugin.frontend.session_options.map((value) => `Session ${plugin.plugin_id}.${value.key}`),
-        ...(plugin.frontend.channel_options ?? []).map((value) => `Channel ${plugin.plugin_id}.${value.key}`),
-        ...plugin.frontend.scheduled_task_options.map((value) => `Scheduled Task ${plugin.plugin_id}.${value.key}`),
-      ]),
-    },
-    {
-      id: "integrations",
-      title: "Integrations",
-      entries: uniqueValues([
-        ...plugin.frontend.tool_renderers.map((value) => `Tool Renderer ${formatToolRendererContribution(value)}`),
-        ...plugin.frontend.file_viewers.map((value) => `File Viewer ${formatFileViewerContribution(value)}`),
-        ...plugin.frontend.skill_importers.map((value) => `Skill Importer ${formatSkillImporterContribution(value)}`),
-        ...plugin.frontend.channel_connectors.map((value) => `Channel Connector ${formatChannelConnectorContribution(value)}`),
-      ]),
-    },
-    {
-      id: "assets-config",
-      title: "Assets And Config",
-      entries: uniqueValues([
-        ...plugin.frontend.settings_sections.map((value) => `Settings Section ${value}`),
-        ...plugin.frontend.i18n_namespaces.map((value) => `I18n ${value}`),
-        ...assetSlots.map((value) => `Asset Slot ${value}`),
-        ...(plugin.package?.data_template.exists
-          ? [plugin.package.layout.data_template || plugin.package.data_template.template || "plugin-data-template"]
-          : []),
-        ...(plugin.package?.layout.has_config_schema ? ["config/schema.json"] : []),
-        ...(plugin.package?.layout.has_config_defaults ? ["config/defaults.json"] : []),
-        ...(plugin.package?.layout.has_resources ? ["resources/resources.yaml"] : []),
-      ]),
-    },
-    {
-      id: "legacy",
-      title: "Legacy Compatibility",
-      entries: legacyFrontendDeclarationLabels(plugin),
-    },
-  ];
-  return groups.filter((group) => group.entries.length > 0);
-}
-
-function emptyContributionSnapshot(): PluginContributionSnapshot {
-  return {
-    appRoutes: [],
-    panels: [],
-    sidebarMoreItems: [],
-    userMenuItems: [],
-    toolRenderers: [],
-    fileViewers: [],
-    skillImporters: [],
-    channelConnectors: [],
-    messageActions: [],
-    chatInputOptions: [],
-    chatInputPanels: [],
-    mentionProviders: [],
-    welcomeSurfaces: [],
-    assistantIdentityResolvers: [],
-    agentCatalogEntries: [],
-    agentCategories: [],
-    projectOptions: [],
-    sessionOptions: [],
-    channelOptions: [],
-    scheduledTaskOptions: [],
-    pluginAssetSlots: [],
-    i18nNamespaces: [],
-  };
 }
 
 function PluginOwnershipOverview({ plugins }: { plugins: PluginRuntimePlugin[] }) {
@@ -408,119 +112,6 @@ function PluginOwnershipOverview({ plugins }: { plugins: PluginRuntimePlugin[] }
       </div>
     </section>
   );
-}
-
-export function buildPluginRuntimeImpactSummary(
-  plugin: PluginRuntimePlugin,
-  runtimePlugins?: PluginRuntimeListResponse["plugins"],
-) {
-  const contributionPreviews = [
-    buildPluginContributionPreview(plugin.plugin_id, runtimePlugins),
-    buildPluginContributionPreview(plugin.plugin_id, runtimePlugins, { agentId: "team" }),
-    buildPluginContributionPreview(plugin.plugin_id, runtimePlugins, { route: "/channels/feishu" }),
-  ];
-  const removedWhenDisabled = contributionPreviews.reduce(
-    (acc: PluginContributionSnapshot, preview): PluginContributionSnapshot => ({
-      appRoutes: uniqueValues([...acc.appRoutes, ...preview.removedWhenDisabled.appRoutes]),
-      panels: uniqueValues([...acc.panels, ...preview.removedWhenDisabled.panels]),
-      sidebarMoreItems: uniqueValues([...acc.sidebarMoreItems, ...preview.removedWhenDisabled.sidebarMoreItems]),
-      userMenuItems: uniqueValues([...acc.userMenuItems, ...preview.removedWhenDisabled.userMenuItems]),
-      toolRenderers: uniqueValues([...acc.toolRenderers, ...preview.removedWhenDisabled.toolRenderers]),
-      fileViewers: uniqueValues([...acc.fileViewers, ...preview.removedWhenDisabled.fileViewers]),
-      skillImporters: uniqueValues([...acc.skillImporters, ...preview.removedWhenDisabled.skillImporters]),
-      channelConnectors: uniqueValues([...acc.channelConnectors, ...preview.removedWhenDisabled.channelConnectors]),
-      messageActions: uniqueValues([...acc.messageActions, ...preview.removedWhenDisabled.messageActions]),
-      chatInputOptions: uniqueValues([...acc.chatInputOptions, ...preview.removedWhenDisabled.chatInputOptions]),
-      chatInputPanels: uniqueValues([...acc.chatInputPanels, ...preview.removedWhenDisabled.chatInputPanels]),
-      mentionProviders: uniqueValues([...acc.mentionProviders, ...preview.removedWhenDisabled.mentionProviders]),
-      welcomeSurfaces: uniqueValues([...acc.welcomeSurfaces, ...preview.removedWhenDisabled.welcomeSurfaces]),
-      assistantIdentityResolvers: uniqueValues([...acc.assistantIdentityResolvers, ...preview.removedWhenDisabled.assistantIdentityResolvers]),
-      agentCatalogEntries: uniqueValues([...acc.agentCatalogEntries, ...preview.removedWhenDisabled.agentCatalogEntries]),
-      agentCategories: uniqueValues([...acc.agentCategories, ...preview.removedWhenDisabled.agentCategories]),
-      projectOptions: uniqueValues([...acc.projectOptions, ...preview.removedWhenDisabled.projectOptions]),
-      sessionOptions: uniqueValues([...acc.sessionOptions, ...preview.removedWhenDisabled.sessionOptions]),
-      channelOptions: uniqueValues([...acc.channelOptions, ...preview.removedWhenDisabled.channelOptions]),
-      scheduledTaskOptions: uniqueValues([...acc.scheduledTaskOptions, ...preview.removedWhenDisabled.scheduledTaskOptions]),
-      pluginAssetSlots: uniqueValues([...acc.pluginAssetSlots, ...preview.removedWhenDisabled.pluginAssetSlots]),
-      i18nNamespaces: uniqueValues([...acc.i18nNamespaces, ...preview.removedWhenDisabled.i18nNamespaces]),
-    }),
-    emptyContributionSnapshot(),
-  );
-  const declaredEntries = declaredRuntimeEntryLabels(plugin);
-
-  return {
-    activeEntries: plugin.executable ? declaredEntries : [],
-    blockedWhenDisabled: uniqueValues([
-      ...declaredEntries,
-      ...removedWhenDisabled.toolRenderers.map(
-        (value) => `renderer ${value}`,
-      ),
-      ...removedWhenDisabled.fileViewers.map(
-        (value) => `viewer ${value}`,
-      ),
-      ...removedWhenDisabled.skillImporters.map(
-        (value) => `importer ${value}`,
-      ),
-      ...removedWhenDisabled.channelConnectors.map(
-        (value) => `connector ${value}`,
-      ),
-      ...removedWhenDisabled.messageActions.map(
-        (value) => `action ${value}`,
-      ),
-      ...removedWhenDisabled.chatInputOptions.map(
-        (value) => `chat option ${value}`,
-      ),
-      ...removedWhenDisabled.chatInputPanels.map(
-        (value) => `chat panel ${value}`,
-      ),
-      ...removedWhenDisabled.mentionProviders.map(
-        (value) => `mention ${value}`,
-      ),
-      ...removedWhenDisabled.welcomeSurfaces.map(
-        (value) => `welcome surface ${value}`,
-      ),
-      ...removedWhenDisabled.assistantIdentityResolvers.map(
-        (value) => `assistant identity ${value}`,
-      ),
-      ...removedWhenDisabled.agentCatalogEntries.map(
-        (value) => `agent ${value}`,
-      ),
-      ...removedWhenDisabled.agentCategories.map(
-        (value) => `agent category ${value}`,
-      ),
-      ...removedWhenDisabled.projectOptions.map(
-        (value) => `project option ${value}`,
-      ),
-      ...removedWhenDisabled.sessionOptions.map(
-        (value) => `session option ${value}`,
-      ),
-      ...removedWhenDisabled.channelOptions.map(
-        (value) => `channel option ${value}`,
-      ),
-      ...removedWhenDisabled.scheduledTaskOptions.map(
-        (value) => `scheduled task option ${value}`,
-      ),
-      ...removedWhenDisabled.pluginAssetSlots.map(
-        (value) => `asset slot ${value}`,
-      ),
-      ...removedWhenDisabled.i18nNamespaces.map(
-        (value) => `i18n ${value}`,
-      ),
-      ...removedWhenDisabled.appRoutes.map(
-        (value) => `app route ${value}`,
-      ),
-      ...removedWhenDisabled.panels.map(
-        (value) => `panel ${value}`,
-      ),
-      ...removedWhenDisabled.sidebarMoreItems.map(
-        (value) => `sidebar ${value}`,
-      ),
-      ...removedWhenDisabled.userMenuItems.map(
-        (value) => `menu ${value}`,
-      ),
-    ]),
-    resourceActions: resourceActionLabels(plugin.dry_run_actions),
-  };
 }
 
 function statusClassName(status: string): string {
