+++ src/api/routes/chat_validation.py
@@ -1,7 +1,7 @@
 """Validation helpers for chat routes."""
 
-from src.kernel.schemas.agent import AgentRequest
 from src.kernel.extensions.plugin_options import plugin_session_options_suppress_core_persona
+from src.kernel.schemas.agent import AgentRequest
 
 
 def validate_team_agent_request(
