+++ src/infra/task/arq_worker.py
@@ -7,8 +7,8 @@ from typing import Any
 from src.agents import ensure_agent_executable
 from src.infra.distributed_validation import validate_distributed_runtime_settings
 from src.infra.logging import get_logger
-from src.kernel.extensions import PluginUnavailableError
 from src.kernel.config import settings
+from src.kernel.extensions import PluginUnavailableError
 
 from .arq_payloads import TaskArqPayloadStore
 from .concurrency import get_concurrency_limiter, get_registered_executor
