+++ src/infra/tool/scheduled_task/helpers.py
@@ -6,9 +6,9 @@ from typing import Any
 from src.infra.logging import get_logger
 from src.infra.role.storage import RoleStorage
 from src.infra.user.storage import UserStorage
+from src.kernel.extensions.plugin_options import selected_agent_team_id_from_metadata
 from src.kernel.schemas.scheduled_task import ChannelDeliveryConfig
 from src.kernel.schemas.user import TokenPayload
-from src.kernel.extensions.plugin_options import selected_agent_team_id_from_metadata
 
 logger = get_logger(__name__)
