+++ src/kernel/extensions/plugin_options.py
@@ -59,9 +59,7 @@ def filter_declared_plugin_options(
         if manifest is None:
             continue
         declared_keys = _declared_option_keys_for_scope(manifest, scope)
-        plugin_values = {
-            key: value for key, value in values.items() if key in declared_keys
-        }
+        plugin_values = {key: value for key, value in values.items() if key in declared_keys}
         if plugin_values:
             filtered[plugin_id] = plugin_values
     return filtered
@@ -258,7 +256,9 @@ def _option_visible_for_agent(option: Any, agent_id: str | None) -> bool:
 def _declared_option_keys_for_scope(manifest: Any, scope: str) -> set[str]:
     declared: set[str] = set()
     declared.update(
-        option.key for option in _frontend_options_for_scope(manifest, scope) if getattr(option, "key", None)
+        option.key
+        for option in _frontend_options_for_scope(manifest, scope)
+        if getattr(option, "key", None)
     )
     declared.update(
         setting.key
@@ -326,7 +326,9 @@ def plugin_agent_ids(
         if getattr(manifest, "id", None) != normalized_plugin_id:
             continue
         agent_ids.extend(
-            agent.id for agent in getattr(manifest, "agents", []) or [] if getattr(agent, "id", None)
+            agent.id
+            for agent in getattr(manifest, "agents", []) or []
+            if getattr(agent, "id", None)
         )
     return agent_ids
 
@@ -349,7 +351,9 @@ def agent_uses_agent_team_options(
     manifests: Any = None,
 ) -> bool:
     """Return whether an agent should consume Agent Team scoped options."""
-    return plugin_id_for_agent(agent_id, runtime=runtime, manifests=manifests) == AGENT_TEAM_PLUGIN_ID
+    return (
+        plugin_id_for_agent(agent_id, runtime=runtime, manifests=manifests) == AGENT_TEAM_PLUGIN_ID
+    )
 
 
 def plugin_session_option_visible_for_agent(
@@ -474,7 +478,13 @@ def with_plugin_option(
 
 
 def _validate_plugin_option_path(*, plugin_id: str, key: str) -> None:
-    if not plugin_id or any(part in {"", ".", ".."} for part in plugin_id.replace("\\", "/").split("/")):
+    if not plugin_id or any(
+        part in {"", ".", ".."} for part in plugin_id.replace("\\", "/").split("/")
+    ):
         raise ValueError("plugin_id must be a safe plugin id")
-    if not key or "." in key or any(part in {"", ".", ".."} for part in key.replace("\\", "/").split("/")):
+    if (
+        not key
+        or "." in key
+        or any(part in {"", ".", ".."} for part in key.replace("\\", "/").split("/"))
+    ):
         raise ValueError("plugin option key must be a safe plugin-local key")
