+++ src/kernel/extensions/registry.py
@@ -9,8 +9,8 @@ from typing import List, Literal
 from src.kernel.extensions.manifest import (
     ExtensionManifest,
     ExtensionType,
-    PluginLifecycleHook,
     PluginAgentCatalogEntry,
+    PluginLifecycleHook,
     PluginManifest,
     PluginRoute,
     PluginTool,
@@ -165,6 +165,7 @@ class ExtensionRegistry:
                     result.append(permission)
         return result
 
+
 class PluginRegistry:
     """Build-time registry for plugin manifests."""
 
@@ -212,9 +213,7 @@ class PluginRegistry:
             if enabled_only and not manifest.enabled_by_default:
                 continue
             for route in manifest.routers:
-                registrations.append(
-                    PluginRouteRegistration(plugin_id=manifest.id, route=route)
-                )
+                registrations.append(PluginRouteRegistration(plugin_id=manifest.id, route=route))
         return registrations
 
     def agents(self, *, enabled_only: bool = True) -> List[PluginAgentRegistration]:
@@ -223,9 +222,7 @@ class PluginRegistry:
             if enabled_only and not manifest.enabled_by_default:
                 continue
             for agent in manifest.agents:
-                registrations.append(
-                    PluginAgentRegistration(plugin_id=manifest.id, agent=agent)
-                )
+                registrations.append(PluginAgentRegistration(plugin_id=manifest.id, agent=agent))
         return registrations
 
     def tools(self, *, enabled_only: bool = True) -> List[PluginToolRegistration]:
@@ -234,9 +231,7 @@ class PluginRegistry:
             if enabled_only and not manifest.enabled_by_default:
                 continue
             for tool in manifest.tools:
-                registrations.append(
-                    PluginToolRegistration(plugin_id=manifest.id, tool=tool)
-                )
+                registrations.append(PluginToolRegistration(plugin_id=manifest.id, tool=tool))
         return registrations
 
     def lifecycle_hooks(
