+++ tests/infra/scheduler/test_runner.py
@@ -12,6 +12,12 @@ import pytest
 
 from src.infra.scheduler.runner import ScheduledTaskRunner
 from src.infra.task.status import TaskStatus
+from src.kernel.extensions import (
+    PluginManifest,
+    PluginRuntime,
+    PluginUnavailableError,
+    build_agent_team_plugin_manifest,
+)
 from src.kernel.schemas.channel import ChannelType
 from src.kernel.schemas.scheduled_task import (
     ChannelDeliveryConfig,
@@ -19,12 +25,6 @@ from src.kernel.schemas.scheduled_task import (
     ScheduledTask,
     ScheduledTaskStatus,
     TriggerType,
-)
-from src.kernel.extensions import (
-    PluginManifest,
-    PluginRuntime,
-    PluginUnavailableError,
-    build_agent_team_plugin_manifest,
 )
 
 
@@ -448,12 +448,14 @@ async def test_execute_agent_hides_injected_timestamp_from_display(
         "source": "scheduled_task",
         "scheduled_task_id": "task_1",
         "scheduled_task_run_id": "run_1",
+        "scheduled_task_trigger_type": "interval",
         "hidden_from_conversation_list": True,
     }
     assert session_manager.metadata == {
         "source": "scheduled_task",
         "scheduled_task_id": "task_1",
         "scheduled_task_run_id": "run_1",
+        "scheduled_task_trigger_type": "interval",
         "hidden_from_conversation_list": True,
     }
 
@@ -616,9 +618,7 @@ async def test_execute_agent_prefers_plugin_option_team_id_for_team_agent(
         input_payload={
             "message": "Plan the launch",
             "team_id": "legacy-team",
-            "plugin_options": {
-                "agent_team": {"SELECTED_TEAM_ID": "plugin-team"}
-            },
+            "plugin_options": {"agent_team": {"SELECTED_TEAM_ID": "plugin-team"}},
         },
     )
     submitted: dict[str, Any] = {}
@@ -847,6 +847,7 @@ async def test_execute_agent_uses_arq_backend_when_enabled(
         "source": "scheduled_task",
         "scheduled_task_id": "task_1",
         "scheduled_task_run_id": "run_1",
+        "scheduled_task_trigger_type": "interval",
         "hidden_from_conversation_list": True,
     }
     assert submitted["write_user_message_immediately"] is True
